WEBAPP-203: Move FailureSwitcher to components
