+++ src/modules/common/components/__tests__/FailureSwitcher.spec.js
